pyproject.toml
[project]
name = "stock-bench"
version = "0.1.0"
description = "LLM marketplace benchmark prototype"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "json5>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[project.scripts]
stock-bench = "stock_bench.cli:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
